refactor: add editor api service and inject dependencies
